+++ .storybook/main.js
@@ -1,25 +1,25 @@
 const path = require("path");
 
 module.exports = {
-  stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
-  addons: [
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-  ],
-  framework: "@storybook/react",
-  core: {
-    builder: "@storybook/builder-webpack5",
-  },
-  webpackFinal: async (config, { configType }) => {
-    // Make whatever fine-grained changes you need
-    config.module.rules.push({
-      test: /\.s[ac]ss$/i,
-      use: ["style-loader", "css-loader", "sass-loader"],
-      include: path.resolve(__dirname, "../"),
-    });
+	stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
+	addons: [
+		"@storybook/addon-links",
+		"@storybook/addon-essentials",
+		"@storybook/addon-interactions"
+	],
+	framework: "@storybook/react",
+	core: {
+		builder: "@storybook/builder-webpack5"
+	},
+	webpackFinal: async (config, { configType }) => {
+		// Make whatever fine-grained changes you need
+		config.module.rules.push({
+			test: /\.s[ac]ss$/i,
+			use: ["style-loader", "css-loader", "sass-loader"],
+			include: path.resolve(__dirname, "../")
+		});
 
-    // Return the altered config
-    return config;
-  },
+		// Return the altered config
+		return config;
+	}
 };